Split last_new_posts_at to last_parsed_at & modified_at
The last_new_posts_at column on rss_feeds was used as a place where the
previous parsing and Last Modified timestamps where stored. The problem
is that Last Modified is published by feed provider, while previousle
parsed is updated by the RSS parsed, so they should be stored in
separate columns.